Decouple factory methods from the provider methods on UserProfileProvider implementation closes #25146
