+++ services/src/main/resources/META-INF/services/org.keycloak.userprofile.UserProfileProviderFactory
@@ -16,4 +16,4 @@
 #  * limitations under the License.
 #  */
 #
-org.keycloak.userprofile.DeclarativeUserProfileProvider
+org.keycloak.userprofile.DeclarativeUserProfileProviderFactory
