+++ testsuite/integration-arquillian/servers/auth-server/services/testsuite-providers/src/main/resources/META-INF/services/org.keycloak.userprofile.UserProfileProviderFactory
@@ -16,4 +16,4 @@
 #  * limitations under the License.
 #  */
 #
-org.keycloak.testsuite.user.profile.CustomUserProfileProvider
+org.keycloak.testsuite.user.profile.CustomUserProfileProviderFactory
